+++ projects/ui-lib-custom/src/lib/theming/index.ts
@@ -3,3 +3,6 @@ export * from './theme-preset.interface';
 export * from './exporters/scss-exporter';
 export * from './exporters/css-exporter';
 export * from './exporters/figma-exporter';
+export * from './theme-scope.directive';
+export * from './theme-directives';
+export * from './with-theme-scope';