+++ projects/ui-lib-custom/theme/src/public-api.ts
@@ -1,2 +1,5 @@
 export * from '../../src/lib/theming/theme-config.service';
 export * from '../../src/lib/theming/theme-preset.interface';
+export * from '../../src/lib/theming/theme-scope.directive';
+export * from '../../src/lib/theming/theme-directives';
+export * from '../../src/lib/theming/with-theme-scope';